Split Indexer implementation from tests
